#101 refactor: 가상 영역 스크롤 되지 않도록 하는 로직 커스텀 훅으로 분리
